updated phase 3

+++ src/main/java/com/haris/notes/oneNote/Service/NoteService.java
@@ -2,7 +2,9 @@ package com.haris.notes.oneNote.Service;
 
 import com.haris.notes.oneNote.Model.ContentObject;
 import com.haris.notes.oneNote.Model.Notes;
+import com.haris.notes.oneNote.Model.User;
 import com.haris.notes.oneNote.Repository.NoteRepository;
+import com.haris.notes.oneNote.Repository.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +17,8 @@ public class NoteService implements IntService {
 
     @Autowired
     NoteRepository noteRepository;
+    @Autowired
+    UserRepo userRepo;
 
 
     @Override
@@ -41,10 +45,13 @@ public class NoteService implements IntService {
     @Override
     public ResponseEntity<?> addNote(ContentObject contentObject, String ownerName) {
 
+            User user = userRepo.findByUsername(ownerName);
+
 
         Notes newnote =  new Notes();
         newnote.setContent(contentObject.getContent());
         newnote.setOwnerName(ownerName);
+        newnote.setUser(user);
         noteRepository.save(newnote);
 
 
@@ -52,37 +59,38 @@ public class NoteService implements IntService {
      }
 
     @Override
-    public ResponseEntity<?> dltNote(int noteid) {
+    public ResponseEntity<?> dltNote(int noteid ,String username) {
 
 
         if (noteRepository.existsById(noteid)) {
 
             for (Notes notes : noteRepository.findAll()) {
-                if (notes.getNoteid() == noteid ) {
+                if (notes.getNoteid() == noteid && notes.getOwnerName().equals(username) ) {
                     noteRepository.delete(notes);
                     return ResponseEntity.ok("Deleted THe Note with id " + noteid);
                 }
             }
 
         }
+
+
+
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Not Found");
 
     }
 
     @Override
-    public ResponseEntity<?> udpateNote(Notes note) {
+    public ResponseEntity<?> udpateNote(Notes note , String username) {
+                for (Notes note1 : noteRepository.findAll()){
+                    if(note1.getNoteid()==note.getNoteid() && note1.getOwnerName().equals(username)){
+                           noteRepository.UpdateNote(note.getNoteid(),note.getOwnerName(),note.getContent());
+                          return ResponseEntity.ok("Update Succesfully");
+                    }
+                }
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Not Updated");
 
-        for (Notes notes : noteRepository.findAll()) {
-            if (notes.getNoteid() != note.getNoteid()) {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Note Not Exits");
-            }
-        }
-        String name = note.getOwnerName();
-        int noteid = note.getNoteid();
-        String notecontent = note.getContent();
 
-         noteRepository.UpdateNote(noteid, name, notecontent);
-         return ResponseEntity.ok(note);
+
 
 
     }
